+++ classes/services/deezer.php
@@ -1,5 +1,34 @@
 <?php
-class WPSSTM_Deezer_Playlists{
+
+class WPSSTM_Deezer{
+    function __construct(){
+        add_action('wpsstm_init_presets',array(__class__,'register_deezer_preset'));
+        add_filter('wpsstm_wizard_services_links',array(__class__,'register_deezer_service_links'));
+    }
+    //register preset
+    static function register_deezer_preset($tracklist){
+        new WPSSTM_Deezer_Preset($tracklist);
+    }
+
+    static function register_deezer_service_links($links){
+        $links[] = array(
+            'slug'      => 'deezer',
+            'name'      => 'Deezer',
+            'url'       => 'https://www.deezer.com',
+            'pages'     => array(
+                array(
+                    'slug'          => 'playlists',
+                    'name'          => __('playlists','wpsstm'),
+                    'example'       => 'http://www.deezer.com/fr/playlist/PLAYLIST_ID',
+                )
+            )
+        );
+
+        return $links;
+    }
+}
+
+class WPSSTM_Deezer_Preset{
     var $tracklist;
     private $playlist_id;
 
@@ -35,27 +64,8 @@ class WPSSTM_Deezer_Playlists{
 
 }
 
-//register preset
-function register_deezer_preset($tracklist){
-    new WPSSTM_Deezer_Playlists($tracklist);
+function wpsstm_deezer_init(){
+    new WPSSTM_Deezer();
 }
 
-function register_deezer_service_links($links){
-    $links[] = array(
-        'slug'      => 'deezer',
-        'name'      => 'Deezer',
-        'url'       => 'https://www.deezer.com',
-        'pages'     => array(
-            array(
-                'slug'          => 'playlists',
-                'name'          => __('playlists','wpsstm'),
-                'example'       => 'http://www.deezer.com/fr/playlist/PLAYLIST_ID',
-            )
-        )
-    );
-
-    return $links;
-}
-
-add_action('wpsstm_get_remote_tracks','register_deezer_preset');
-add_filter('wpsstm_wizard_services_links','register_deezer_service_links');
+add_action('wpsstm_init','wpsstm_deezer_init');
