+++ classes/wpsstm-live-tracklist-class.php
@@ -132,7 +132,7 @@ class WPSSTM_Remote_Tracklist extends WPSSTM_Tracklist{
         }
         
         $is_cached = false;
-        do_action('wpsstm_get_remote_tracks',$this);
+        do_action('wpsstm_init_presets',$this);
 
         //try cache
         if ( $this->get_options('cache_source') && $this->cache_source_url ){
